requirements.txt
mysql-connector-python==8.0.17
python-decouple==3.1
